:recycle: UserService register 셀러 셀러 등록 체크 부분 및 셀러 저장 메서드 결합도 분리
